# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GrabGo API
  description: |
    Backend API for GrabGo food delivery platform.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## User Roles
    - **customer**: Regular users who order food
    - **restaurant**: Restaurant owners/managers
    - **rider**: Delivery riders
    - **admin**: Platform administrators
  version: 1.0.0
  contact:
    name: GrabGo Support
  license:
    name: ISC
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.grabgo.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/register endpoints
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
...
    ScheduledNotification:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        scheduledFor:
          type: string
          format: date-time
          description: When to send the notification
        timezone:
          type: string
          default: UTC
          example: "Africa/Accra"
        status:
          type: string
          enum: [pending, sent, cancelled, failed]
          description: Current status of the scheduled notification
        type:
          type: string
          enum: [order, promo, update, system, comment_reply, comment_reaction, referral_completed, payment_confirmed, delivery_arriving, milestone_bonus]
          description: Type of notification
        title:
          type: string
          maxLength: 200
          example: "Flash Sale!"
        message:
          type: string
          maxLength: 500
          example: "Get 50% off all orders for the next hour!"
        data:
          type: object
          description: Additional data for deep linking
          properties:
            route:
              type: string
            promoCode:
              type: string
        targetType:
          type: string
          enum: [user, segment, all]
          description: Who should receive this notification
        targetUsers:
          type: array
          items:
            type: string
          description: Specific user IDs (when targetType is 'user')
        targetSegment:
          type: object
          description: Segment criteria (when targetType is 'segment')
          properties:
            inactive:
              type: boolean
            daysSinceLastOrder:
              type: integer
            newUsers:
              type: boolean
            registeredWithinDays:
              type: integer
        isRecurring:
          type: boolean
          default: false
        recurrencePattern:
          type: object
          properties:
            frequency:
              type: string
              enum: [daily, weekly, monthly]
            timeOfDay:
              type: string
              example: "12:00"
            endDate:
              type: string
              format: date-time
        createdBy:
          type: string
          description: Admin user who created this
        sentAt:
          type: string
          format: date-time
        failureReason:
          type: string
        retryCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Notification:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
        type:
          type: string
          enum: [order, promo, update, system, comment_reply, comment_reaction, referral_completed, payment_confirmed, delivery_arriving, milestone_bonus]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        data:
          type: object
        actors:
          type: array
          items:
            type: object
            properties:
              actorId:
                type: string
              actorName:
                type: string
              actorAvatar:
                type: string
        actorCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    NotificationSettings:
      type: object
      properties:
        chatMessages:
          type: boolean
          default: true
        orderUpdates:
          type: boolean
          default: true
        promoNotifications:
          type: boolean
          default: true
        commentReplies:
          type: boolean
          default: true
        commentReactions:
          type: boolean
          default: true
        referralUpdates:
          type: boolean
          default: true
        paymentUpdates:
          type: boolean
          default: true
        deliveryUpdates:
          type: boolean
          default: true
        systemUpdates:
          type: boolean
          default: true
paths:
...
  /scheduled-notifications:
    post:
      tags: [Scheduled Notifications]
      summary: Create a scheduled notification
      description: Schedule a notification to be sent at a specific future time (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduledFor, type, title, message]
              properties:
                scheduledFor:
                  type: string
                  format: date-time
                  example: "2025-12-25T09:00:00Z"
                timezone:
                  type: string
                  default: UTC
                type:
                  type: string
                  enum: [order, promo, update, system, comment_reply, comment_reaction, referral_completed, payment_confirmed, delivery_arriving, milestone_bonus]
                title:
                  type: string
                  maxLength: 200
                  example: "Christmas Special!"
                message:
                  type: string
                  maxLength: 500
                  example: "Get 50% off today!"
                data:
                  type: object
                  example: {"promoCode": "XMAS50", "route": "/promos"}
                targetType:
                  type: string
                  enum: [user, segment, all]
                  default: all
                targetUsers:
                  type: array
                  items:
                    type: string
                targetSegment:
                  type: object
                isRecurring:
                  type: boolean
                  default: false
                recurrencePattern:
                  type: object
                  properties:
                    frequency:
                      type: string
                      enum: [daily, weekly, monthly]
                    timeOfDay:
                      type: string
                    endDate:
                      type: string
                      format: date-time
      responses:
        '201':
          description: Scheduled notification created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  scheduledNotification:
                    $ref: '#/components/schemas/ScheduledNotification'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

    get:
      tags: [Scheduled Notifications]
      summary: Get all scheduled notifications
      description: List scheduled notifications with optional filters (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, sent, cancelled, failed]
        - name: type
          in: query
          schema:
            type: string
            enum: [order, promo, update, system]
        - name: scheduledAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: scheduledBefore
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Scheduled notifications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduledNotification'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
  /scheduled-notifications/stats:
    get:
      tags: [Scheduled Notifications]
      summary: Get scheduled notification statistics
      description: Get counts by status (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  stats:
                    type: object
                    properties:
                      pending:
                        type: integer
                      sent:
                        type: integer
                      cancelled:
                        type: integer
                      failed:
                        type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
  /scheduled-notifications/{id}:
    get:
      tags: [Scheduled Notifications]
      summary: Get single scheduled notification
      description: Get detailed information about a specific scheduled notification (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scheduled notification retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  notification:
                    $ref: '#/components/schemas/ScheduledNotification'
        '404':
          description: Scheduled notification not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

    patch:
      tags: [Scheduled Notifications]
      summary: Update a scheduled notification
      description: Update a pending scheduled notification (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduledFor:
                  type: string
                  format: date-time
                title:
                  type: string
                  maxLength: 200
                message:
                  type: string
                  maxLength: 500
                data:
                  type: object
                targetType:
                  type: string
                  enum: [user, segment, all]
                targetUsers:
                  type: array
                  items:
                    type: string
                targetSegment:
                  type: object
      responses:
        '200':
          description: Scheduled notification updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  notification:
                    $ref: '#/components/schemas/ScheduledNotification'
        '400':
          description: Validation error
        '404':
          description: Scheduled notification not found or cannot be updated
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

    delete:
      tags: [Scheduled Notifications]
      summary: Cancel a scheduled notification
      description: Cancel a pending scheduled notification (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scheduled notification cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          description: Scheduled notification not found or cannot be cancelled
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

  # ============ NOTIFICATIONS ============
  /notifications:
    get:
      tags: [Notifications]
      summary: Get user notifications
      description: Get paginated list of user's notifications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      hasMore:
                        type: boolean
                      totalPages:
                        type: integer
...
  /users/settings/notifications:
    get:
      tags: [Notification Settings]
      summary: Get user's notification settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  notificationSettings:
                    $ref: '#/components/schemas/NotificationSettings'

    patch:
      tags: [Notification Settings]
      summary: Update user's notification settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings:
                  type: object
                  properties:
                    chatMessages:
                      type: boolean
                    orderUpdates:
                      type: boolean
                    promoNotifications:
                      type: boolean
                    commentReplies:
                      type: boolean
                    commentReactions:
                      type: boolean
                    referralUpdates:
                      type: boolean
                    paymentUpdates:
                      type: boolean
                    deliveryUpdates:
                      type: boolean
                    systemUpdates:
                      type: boolean
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  notificationSettings:
                    $ref: '#/components/schemas/NotificationSettings'
        '400':
          description: Invalid settings format

  # ============ HEALTH ============
...